+++ operators/concatenate_strips.py
@@ -30,31 +30,22 @@ class POWER_SEQUENCER_OT_concatenate_strips(bpy.types.Operator):
         "shortcuts": [
             (
                 {"type": "C", "value": "PRESS"},
-                {"concatenate_all": False, "direction": "left"},
-                (
-                    "Concatenate selected strips in channel, or"
-                    " concatenate & select next strip in channel if only"
-                    " 1 strip selected"
-                ),
+                {"concatenate_all": False, "to_left": True},
+                ("Concatenate and select the next strip in the channel"),
             ),
             (
                 {"type": "C", "value": "PRESS", "shift": True},
-                {"concatenate_all": True, "direction": "left"},
+                {"concatenate_all": True, "to_left": True},
                 "Concatenate all strips in selected channels",
             ),
             (
                 {"type": "C", "value": "PRESS", "alt": True},
-                {"concatenate_all": False, "direction": "right"},
-                (
-                    "Concatenate selected strips in channel"
-                    " towards the right, or concatenate and select"
-                    " the previous strip in the channel if only 1"
-                    " strip selected"
-                ),
+                {"concatenate_all": False, "to_left": False},
+                ("Concatenate and select the previous strip in the channel towards the right"),
             ),
             (
                 {"type": "C", "value": "PRESS", "shift": True, "alt": True},
-                {"concatenate_all": True, "direction": "right"},
+                {"concatenate_all": True, "to_left": False},
                 "Shift Alt C; Concatenate all strips in channel towards the right",
             ),
         ],
@@ -65,21 +56,15 @@ class POWER_SEQUENCER_OT_concatenate_strips(bpy.types.Operator):
     bl_description = doc_brief(doc["description"])
     bl_options = {"REGISTER", "UNDO"}
 
-    frame, channel = -1, -1
-
     concatenate_all: bpy.props.BoolProperty(
         name="Concatenate all strips in channel",
         description=("If only one strip selected, concatenate" " the entire channel"),
         default=False,
     )
-    direction: bpy.props.EnumProperty(
-        name="Direction",
-        description=("Concatenate strips moving them back in time (default)" " or forward in time"),
-        items=[
-            ("left", "Left", "Move strips back in time, to the left"),
-            ("right", "Right", "Move strips forward in time, to the right"),
-        ],
-        default="left",
+    to_left: bpy.props.BoolProperty(
+        name="To Left",
+        description="Concatenate strips moving them back in time (default) or forward in time",
+        default=True,
     )
 
     @classmethod
@@ -87,105 +72,66 @@ class POWER_SEQUENCER_OT_concatenate_strips(bpy.types.Operator):
         return context.sequences
 
     def invoke(self, context, event):
-        frame, channel = get_mouse_frame_and_channel(context, event)
         if not context.selected_sequences:
+            frame, channel = get_mouse_frame_and_channel(context, event)
             bpy.ops.power_sequencer.select_closest_to_mouse(frame=frame, channel=channel)
         return self.execute(context)
 
     def execute(self, context):
         selection = context.selected_sequences
-        one_strip_only = True if len(selection) == 1 else False
         channels = {s.channel for s in selection}
 
-        # If only one strip selected per channel,
-        # Loop over each strip and detect which strips to concatenate
-        if len(channels) == len(selection):
+        if len(selection) == len(channels):
             for s in selection:
-                if self.direction == "right":
-                    in_channel = [
-                        strip
-                        for strip in find_sequences_before(context, s)
-                        if strip.channel == s.channel
-                    ]
-                else:
-                    in_channel = [
-                        strip
-                        for strip in find_sequences_after(context, s)
-                        if strip.channel == s.channel
-                    ]
-                in_channel.append(s)
+                candidates = (
+                    find_sequences_before(context, s)
+                    if not self.to_left
+                    else find_sequences_after(context, s)
+                )
                 to_concatenate = [
-                    strip for strip in in_channel if strip.type in SequenceTypes.CUTABLE
+                    strip
+                    for strip in candidates
+                    if strip.channel == s.channel
+                    and not strip.lock
+                    and strip.type in SequenceTypes.CUTABLE
                 ]
+                self.concatenate(s, to_concatenate)
 
-                if self.direction == "right":
-                    self.concatenate_right(to_concatenate, one_strip_only)
-                else:
-                    self.concatenate_left(to_concatenate, one_strip_only)
         else:
             for channel in channels:
                 to_concatenate = [s for s in selection if s.channel == channel]
-                if self.direction == "right":
-                    self.concatenate_right(to_concatenate)
-                else:
-                    self.concatenate_left(to_concatenate)
+                strip_target = (
+                    min(to_concatenate, key=lambda s: s.frame_final_start)
+                    if self.to_left
+                    else max(to_concatenate, key=lambda s: s.frame_final_start)
+                )
+                to_concatenate.remove(strip_target)
+                self.concatenate(strip_target, to_concatenate, force_all=True)
         return {"FINISHED"}
 
-    def concatenate_left(self, sequences, one_strip_only=False):
-        """
-        Takes a list of sequences in a single channel, sorts them by frame_final_start,
-        and concatenates them.
-        """
-        if len(sequences) <= 1:
-            return
-        sorted_sequences = sorted(sequences, key=attrgetter("frame_final_start"))
-        first_strip = sorted_sequences[0]
-        if self.concatenate_all or not one_strip_only:
-            to_concatenate = sorted_sequences[1:]
-        else:
-            first_strip.select = False
-            second_strip = sorted_sequences[1]
-            second_strip.select = True
-            to_concatenate = [second_strip]
+    def concatenate(self, strip_target, sequences, force_all=False):
+        to_concatenate = sorted(sequences, key=attrgetter("frame_final_start"))
+        to_concatenate = list(reversed(to_concatenate)) if not self.to_left else to_concatenate
+        to_concatenate = (
+            [to_concatenate[0]] if not (self.concatenate_all or force_all) else to_concatenate
+        )
 
-        concatenate_start = first_strip.frame_final_end
+        attribute_target = "frame_final_end" if self.to_left else "frame_final_start"
+        attribute_concat = "frame_final_start" if self.to_left else "frame_final_end"
+        concatenate_start = getattr(strip_target, attribute_target)
         last_gap = 0
         for s in to_concatenate:
-            gap = s.frame_final_start - concatenate_start
             if isinstance(s, bpy.types.EffectSequence):
-                concatenate_start = s.frame_final_end - last_gap
+                concatenate_start = (
+                    s.frame_final_end - last_gap if self.to_left else s.frame_final_start - last_gap
+                )
                 continue
-            last_gap = gap
+            concat_strip_frame = getattr(s, attribute_concat)
+            gap = concat_strip_frame - concatenate_start
             s.frame_start -= gap
-            concatenate_start = s.frame_final_end
-
-    def concatenate_right(self, sequences, one_strip_only=False):
-        """
-        Takes a list of sequences in a single channel, sorts them by frame_final_start,
-        and concatenates them moving strips forward in time, towards the last strip in
-        the ordered list
-        """
-        if len(sequences) <= 1:
-            return
-        sorted_sequences = sorted(sequences, key=attrgetter("frame_final_start"))
-        last_strip = sorted_sequences.pop()
-        if self.concatenate_all or not one_strip_only:
-            to_concatenate = sorted_sequences
-        else:
-            last_strip.select = False
-            second_strip = sorted_sequences.pop()
-            second_strip.select = True
-            to_concatenate = [second_strip]
-        print(to_concatenate)
-
-        concatenate_start = last_strip.frame_final_start
-        last_gap = 0
-        print(concatenate_start)
-        for s in reversed(to_concatenate):
-            gap = s.frame_final_end - concatenate_start
-            if isinstance(s, bpy.types.EffectSequence):
-                concatenate_start = s.frame_final_start - last_gap
-                continue
+            concatenate_start = s.frame_final_end if self.to_left else s.frame_final_start
             last_gap = gap
-            s.frame_start -= gap
-            concatenate_start = s.frame_final_start
+
+        if not self.concatenate_all:
+            strip_target.select = False
+            to_concatenate[0].select = True