+++ operators/utils/functions.py
@@ -85,8 +85,7 @@ def find_sequences_after(context, sequence):
     - Sequences, the sequences to check
     Returns all the strips after the sequence in the current context
     """
-    sequence_start = sequence.frame_final_start
-    return [s for s in context.sequences if s.frame_final_start > sequence_start]
+    return [s for s in context.sequences if s.frame_final_start > sequence.frame_final_start]
 
 
 def find_snap_candidate(context, frame=0):